requirements.txt
python-dotenv
langchain
langchain-chroma
langchain-community
langchain-groq
grandalf
numpy
pandas
sentence-transformers
langchain-core
langchain-openai
langchain-anthropic
tiktoken 
langchainhub
langgraph
llama-index
llama-index-core
llama-parse
openai
llama-index-llms-anthropic
llama_index.embeddings.huggingface
giskard
langchain-core
ragas
pyarrow

